update: restructure src/core
